requirements.txt
PyQt5>=5.15.0
python-dotenv>=1.0.0
requests>=2.28.0
openai>=1.0.0
pillow>=9.0.0
pywin32>=300; sys_platform == 'win32'

# 向量数据库相关依赖
numpy>=1.21.0
tqdm>=4.64.0